// packages/satteri/src/mdx-types.ts
export interface MdxJsxAttributeValueExpression extends Node {
  type: "mdxJsxAttributeValueExpression";
  value: string;
  data?: MdxJsxAttributeValueExpressionData | undefined;
}

export interface MdxJsxAttributeValueExpressionData extends Data {}

export interface MdxJsxExpressionAttribute extends Node {
  type: "mdxJsxExpressionAttribute";
  value: string;
  data?: MdxJsxExpressionAttributeData | undefined;
}

export interface MdxJsxExpressionAttributeData extends Data {}

export interface MdxJsxAttribute extends Node {
  type: "mdxJsxAttribute";
  name: string;
  value?: MdxJsxAttributeValueExpression | string | null | undefined;
  data?: MdxJsxAttributeData | undefined;
}

export interface MdxJsxAttributeData extends Data {}

export interface MdxJsxFlowElement extends MdastParent {
  type: "mdxJsxFlowElement";
  name: string | null;
  attributes: Array<MdxJsxAttribute | MdxJsxExpressionAttribute>;
  children: Array<BlockContent | DefinitionContent>;
  data?: MdxJsxFlowElementData | undefined;
}

export interface MdxJsxFlowElementData extends MdastData {}